requirements.txt
openai==1.78.1
streamlit==1.45.0
python-dotenv==1.1.0
